challenges 4.20.-4.21.-4.22 completed

+++ controllers/blogs.js
@@ -12,21 +12,14 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs);
 });
 
-// const getTokenFrom = (request) => {
-//   const authorization = request.get('authorization');
-//   if (authorization && authorization.startsWith('Bearer ')) {
-//     return authorization.replace('Bearer ', '');
-//   }
-//   return null;
-// };
-
 blogsRouter.post('/', async (request, response) => {
   const body = request.body;
-  const decodedToken = jwt.verify(request.token, process.env.SECRET);
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' });
-  }
-  const user = await User.findById(decodedToken.id);
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' });
+  // }
+  // const user = await User.findById(decodedToken.id);
+  const user = request.user;
 
   const blog = new Blog({
     title: body.title,
@@ -41,6 +34,8 @@ blogsRouter.post('/', async (request, response) => {
   }
 
   const savedBlog = await blog.save();
+  console.log(savedBlog);
+  console.log(savedBlog._id);
   user.blogs = user.blogs.concat(savedBlog._id);
   await user.save();
 
@@ -48,8 +43,28 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' });
+  // }
+
+  // const user = await User.findById(decodedToken.id);
+  const user = request.user;
+  const blog = await Blog.findById(request.params.id);
+  console.log(user);
+  console.log(blog);
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: 'invalid user' });
+  }
+  user.blogs = user.blogs.filter(
+    (userBlog) => userBlog.toString() !== blog._id.toString()
+  );
+  console.log(user.blogs);
+  await user.save();
+  // user.blogs = correctedBlogs;
 });
 
 blogsRouter.put('/:id', async (request, response) => {
